+++ src/leveldb/util/logging.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "util/logging.h"
+#include "leveldb/util/logging.h"
 
 #include <errno.h>
 #include <stdarg.h>
@@ -64,7 +64,7 @@ bool ConsumeDecimalNumber(Slice* in, uint64_t* val) {
       const int delta = (c - '0');
       static const uint64_t kMaxUint64 = ~static_cast<uint64_t>(0);
       if (v > kMaxUint64/10 ||
-          (v == kMaxUint64/10 && delta > kMaxUint64%10)) {
+          (v == kMaxUint64/10 && (uint64_t)delta > kMaxUint64%10)) {
         // Overflow
         return false;
       }
